+++ utils.h
@@ -1,6 +1,7 @@
 #ifndef _UTILS_H
 #define _UTILS_H
 
-char* itoa(int num, char* str, int radix);
+#include <linux/types.h>
+char* itoa(int64_t num, char* str, int radix);
 
 #endif
